requirements.txt
google-cloud==0.34.0
google-cloud-speech==0.35.0
pyaudio==0.2.10
wave==0.0.2
requests==2.18.4